renamed files; added test tupfile
