+++ hatch/src/dtl/tup/mod.rs
@@ -1,6 +1,7 @@
-pub mod build_system;
 pub mod config;
+pub mod tupfile;
 pub mod platform;
+pub mod tuprules;
 pub mod test;
 
 use std::fmt;
