+++ hatch/src/dtl/tup/test.rs
@@ -1,27 +1,15 @@
-use project::Project;
-use dtl::tup::Asset;
+use project::{ LibraryKind, ProjectKind };
 
-pub struct TestAssets {
-  file_path: String,
-  file_contents: String,
-}
+pub struct Tupfile;
 
-impl Asset for TestAssets {
-  fn path(&self) -> &str {
-    &self.file_path.as_str()
-  }
-
-  fn contents(&self) -> &str {
-    &self.file_contents.as_str()
+impl Tupfile {
+  pub fn new() -> Tupfile {
+    Tupfile
   }
 }
 
-impl TestAssets {
-  pub fn tupfile(project: &Project) -> TestAssets {
-    let file_path = "C++/libs/".to_owned() + project.name() + "/test/Tupfile";
-    let file_contents =
-".gitignore".to_string();
-
-    TestAssets { file_path, file_contents }
+impl ToString for Tupfile {
+  fn to_string(&self) -> String {
+    String::from(".gitignore")
   }
 }
